Extract eventNodesDispoising module
